requirements.txt
# Core ML libraries
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# MLOps tools
dvc[all]==3.15.2
pyyaml==6.0

# Model persistence
joblib==1.3.2

# Development tools
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Optional: Jupyter for experimentation
# jupyter==1.0.0
# matplotlib==3.7.2
# seaborn==0.12.2
